Generate entry timestamps using correct relative point in time
The intersection observer polyfill and native intersection observers use different timestamps. The polyfill's timestamps are relative to 00:00:00 UTC on 1 January 1970, whereas the native io uses highres timestamps which are relative to navigation start. I had forgotten that we mimick generating an observer entry when re-showing a tab, and this code didn't respect the differences in timestamps.