Fix test issue with collecting initial page load events
When a webpage boots and elements are instrumented with a watcher, the watcher will fire exposed events during page boot for elements that are above the fold. Firing these exposed events are not instantaneous and our tests were not waiting enough to capture all of them. Adding some grace period to wait until these events are collected.