Better typing for spaniel observer payload
Previously, the spaniel observer entry payload was typed with "any." This is obviously bad, and I've updated the code to use a generic instead, so the payload is typed. Also, we were defining label as optional in the spaniel observer entry, when it is guaranteed to exist since we have always required the label in the threshold that users provide.